requirements.txt
meshtastic
paho-mqtt
peewee
protobuf
psutil
pubsub